farplug-wide.h: StructSize is size_t in all structs (was: int)

+++ far/far2sdk/farplug-wide.h
@@ -584,7 +584,7 @@ struct OpenDlgPluginData
 
 struct DialogInfo
 {
-	int StructSize;
+	size_t StructSize;
 	GUID Id;
 	DWORD Owner;
 };
@@ -1709,7 +1709,7 @@ struct ViewerMode
 
 struct ViewerInfo
 {
-	int    StructSize;
+	size_t StructSize;
 	int    ViewerID;
 	const wchar_t *FileName;
 	int64_t FileSize;
@@ -2328,7 +2328,7 @@ enum BOX_DEF_SYMBOLS
 
 typedef struct FarStandardFunctions
 {
-	int StructSize;
+	size_t StructSize;
 
 	FARSTDATOI                 atoi;
 	FARSTDATOI64               atoi64;
@@ -2403,7 +2403,7 @@ typedef struct FarStandardFunctions
 
 struct PluginStartupInfo
 {
-	int StructSize;
+	size_t StructSize;
 	const wchar_t *ModuleName;
 	INT_PTR ModuleNumber;
 	const wchar_t *RootKey;
@@ -2468,7 +2468,7 @@ enum PLUGIN_FLAGS
 
 struct PluginInfo
 {
-	int StructSize;
+	size_t StructSize;
 	DWORD Flags;
 	const wchar_t * const *DiskMenuStrings;
 	int *Reserved0;
@@ -2579,7 +2579,7 @@ enum OPERATION_MODES
 
 struct OpenPluginInfo
 {
-	int                           StructSize;
+	size_t                        StructSize;
 	DWORD                         Flags;
 	const wchar_t*                HostFile;
 	const wchar_t*                CurDir;
