+++ far/src/dialog.cpp
@@ -4761,14 +4761,14 @@ LONG_PTR WINAPI DefDlgProc(HANDLE hDlg,int Msg,int Param1,LONG_PTR Param2)
 				DialogInfo *di=reinterpret_cast<DialogInfo*>(Param2);
 				if (Dlg->IdExist)
 				{
-					if (static_cast<size_t>(di->StructSize)>=offsetof(DialogInfo,Id)+sizeof(di->Id))
+					if (di->StructSize >= offsetof(DialogInfo,Id)+sizeof(di->Id))
 					{
 						di->Id=Dlg->Id;
 						Result=TRUE;
 					}
 				}
 
-				if (static_cast<size_t>(di->StructSize)>=offsetof(DialogInfo,Owner)+sizeof(di->Owner))
+				if (di->StructSize >= offsetof(DialogInfo,Owner)+sizeof(di->Owner))
 				{
 					di->Owner = 0;
 					if (Dlg->PluginNumber != -1)
